+++ internal/mesh/peer.go
@@ -29,7 +29,6 @@ type Peer struct {
 	Endpoint            string   `codec:"ep"`
 	AllowedIPs          []string `codec:"ai"`
 	PersistentKeepalive int      `codec:"k,omitempty"`
-	Tags                Tags     `codec:"t,omitempty"`
 	Signature           []byte   `codec:"s,omitempty"`
 }